feat: add jump from init to inportusers in mis-web
